refactor drivers as rides
