+++ client/apis/drivers.js
@@ -17,18 +17,6 @@ const rootUrl = '/api/v1'
 //   return await request.post(rootUrl + '/drivers/')
 //     .send(ride)
 // }
-export function getRides () {
-  return request.get(rootUrl + '/drivers/')
-    .then(res => {
-      return res.body.rides
-    })
-}
-
-export async function addRides (ride, user) {
-  const rideUser = { newRide: ride, auth: user }
-  return await request.post(rootUrl + '/drivers/')
-    .send(rideUser)
-}
 
 export function getCar () {
   return request.get(rootUrl + '/drivers')
