+++ client/components/Drivers/PostRide.jsx
@@ -11,7 +11,7 @@ import FormControl from '@mui/material/FormControl'
 import MenuItem from '@mui/material/MenuItem'
 import store from '../../store'
 
-import { addRides } from '../../apis/drivers'
+import { addRides } from '../../apis/rides'
 
 const rideSchema = Yup.object().shape({
   startLocation: Yup.string()
