+++ server/routes/rides.js
@@ -6,7 +6,9 @@ const dbRides = require('../db/rides')
 
 const router = express.Router()
 
-// ROUTE api/v1/drivers/
+// ROUTE api/v1/rides/
+
+// this get route is for the search bar:
 
 router.get('/', (req, res) => {
   console.log('query string params: ', req.query)
