+++ server/server.js
@@ -5,7 +5,7 @@ const publicRoutes = require('./routes/public')
 const protectedRoutes = require('./routes/protected')
 const privateRoutes = require('./routes/private')
 const usersRoutes = require('./routes/users')
-const driverRoutes = require('./routes/drivers')
+const ridesRoutes = require('./routes/rides')
 
 const server = express()
 
@@ -16,6 +16,6 @@ server.use('/api/v1/users', usersRoutes)
 server.use('/api/v1/public', publicRoutes)
 server.use('/api/v1/protected', protectedRoutes)
 server.use('/api/v1/private', privateRoutes)
-server.use('/api/v1/drivers', driverRoutes)
+server.use('/api/v1/rides', ridesRoutes)
 
 module.exports = server
